requirements/base.txt
dj-database-url==0.5.0
dj-static==0.0.6
django-cors-headers==2.3.0
Django==2.0.7
djangorestframework-jwt==1.11.0
djangorestframework==3.8.2
psycopg2-binary==2.7.5
pytz==2018.5
